Refactor pages into modular components for better code organization
