+++ app/_components/HeroSection.tsx
@@ -1,0 +1,10 @@
+export default function HeroSection() {
+  return (
+    <div className="px-4 pt-8 pb-6">
+      <div className="text-center mb-8">
+        <h1 className="text-3xl font-bold text-gray-900 mb-2">定点撮影SNS</h1>
+        <p className="text-gray-600">同じ場所から時の流れを記録しよう</p>
+      </div>
+    </div>
+  );
+}
