+++ app/capture/_components/CaptureHeader.tsx
@@ -1,0 +1,9 @@
+export default function CaptureHeader() {
+  return (
+    <div className="absolute top-0 left-0 right-0 z-10 flex items-center justify-between p-4 bg-gradient-to-b from-black/50 to-transparent">
+      <div className="w-10" />
+      <h1 className="text-white font-medium">撮影ガイド</h1>
+      <div className="w-10" />
+    </div>
+  );
+}
