+++ app/register-location/_components/LocationHeader.tsx
@@ -1,0 +1,8 @@
+export default function LocationHeader() {
+  return (
+    <div className="flex items-center gap-3 p-4 border-b">
+      <div className="w-10" />
+      <h1 className="text-lg font-semibold">撮影地点を登録</h1>
+    </div>
+  );
+}
