+++ app/timeline/_components/TimelineHeader.tsx
@@ -1,0 +1,7 @@
+export default function TimelineHeader() {
+  return (
+    <div className="bg-white border-b p-4">
+      <h1 className="text-xl font-bold">タイムライン</h1>
+    </div>
+  );
+}
